+++ src/index.ts
@@ -1,0 +1,37 @@
+export { createFiroRpcClient } from './utils/client.js';
+export { RpcCallError } from './utils/rpc.js';
+export type { FiroRpcClient } from './utils/client.js';
+
+export type {
+  // RPC core
+  RpcConfig,
+  RpcRequest,
+  RpcResponse,
+  RpcError,
+  BatchCall,
+  BatchResult,
+  // Firo types
+  Block,
+  BlockchainInfo,
+  Bip9SoftFork,
+  SoftFork,
+  Transaction,
+  TxType,
+  Vin,
+  CoinbaseVin,
+  StandardVin,
+  SparkSpendVin,
+  Vout,
+  ScriptSig,
+  ScriptPubKey,
+  CbTx,
+  ProUpServTx,
+  FinalCommitment,
+  QuorumCommitment,
+  TxOutSetInfo,
+  NetworkInfo,
+  PeerInfo,
+  MempoolInfo,
+  MempoolEntry,
+  RawMempool,
+} from './types/index.js';