+++ src/types/block.ts
@@ -1,0 +1,24 @@
+import type { CbTx, Transaction } from './transactions.js';
+
+export interface Block {
+  hash: string;
+  confirmations: number;
+  strippedsize: number;
+  size: number;
+  weight: number;
+  height: number;
+  version: number;
+  versionHex: string;
+  merkleroot: string;
+  tx: Transaction[];
+  cbTx: CbTx;
+  time: number;
+  mediantime: number;
+  nonce: number;
+  bits: string;
+  difficulty: number;
+  chainwork: string;
+  previousblockhash: string;
+  nextblockhash?: string;
+  chainlock: boolean;
+}