+++ src/types/index.ts
@@ -1,0 +1,32 @@
+export type {
+  RpcConfig,
+  RpcRequest,
+  RpcResponse,
+  RpcError,
+  BatchCall,
+  BatchResult,
+} from './rpc.js';
+
+export type { Block } from './block.js';
+
+export type { BlockchainInfo, Bip9SoftFork, SoftFork, TxOutSetInfo } from './blockchain.js';
+
+export type {
+  Transaction,
+  TxType,
+  Vin,
+  CoinbaseVin,
+  StandardVin,
+  SparkSpendVin,
+  Vout,
+  CbTx,
+  ProUpServTx,
+  FinalCommitment,
+  QuorumCommitment,
+} from './transactions.js';
+
+export type { ScriptSig, ScriptPubKey } from './shared.js';
+
+export type { NetworkInfo, PeerInfo } from './network.js';
+
+export type { MempoolInfo, MempoolEntry, RawMempool } from './mempool.js';