+++ src/types/network.ts
@@ -1,0 +1,52 @@
+export interface NetworkInfo {
+  version: number;
+  subversion: string;
+  protocolversion: number;
+  localservices: string;
+  localrelay: boolean;
+  timeoffset: number;
+  networkactive: boolean;
+  connections: number;
+  networks: {
+    name: string;
+    limited: boolean;
+    reachable: boolean;
+    proxy: string;
+    proxy_randomize_credentials: boolean;
+  }[];
+  relayfee: number;
+  incrementalfee: number;
+  localaddresses: string[];
+  warnings: string;
+}
+
+export interface PeerInfo {
+  id: number;
+  addr: string;
+  addrlocal: string;
+  services: string;
+  verified_proregtx_hash?: string;
+  relaytxes: boolean;
+  lastsend: number;
+  lastrecv: number;
+  bytessent: number;
+  bytesrecv: number;
+  conntime: number;
+  timeoffset: number;
+  pingtime: number;
+  minping: number;
+  version: number;
+  subver: string;
+  inbound: boolean;
+  addnode: boolean;
+  startingheight: number;
+  banscore: number;
+  synced_headers: number;
+  synced_blocks: number;
+  inflight: number[];
+  addr_processed: number;
+  addr_rate_limited: number;
+  whitelisted: number;
+  bytessent_per_msg: Record<string, number>;
+  bytesrecv_per_msg: Record<string, number>;
+}